New cp formula (#85)
Bench: 3372243

+++ engine/src/search_engine/engine_options.rs
@@ -57,8 +57,8 @@ create_options! {
             visit_increase_offset:   f64  =>  4.1092,  1.0,  10.0,   0.1,   0.002;
 
             //Butterfly history
-            butterfly_reduction_factor: i64 => 8192,   1,  65536,   819,   0.002;
-            butterfly_bonus_scale:      i64 => 16384,  1,  131072,  1638,  0.002;
+            butterfly_reduction_factor: i64 => 8192,     1,    65536,     819,     0.002;
+            butterfly_bonus_scale:      f64 => 16384.0,  1.0,  131072.0,  1638.0,  0.002;
 
             //Draw Scaling
             power_50mr:          f64  =>  3.0,     1.0,  10.0,  0.3,     0.002;
